Replace XVisualInfo usage by GLXFBConfig in order to fix rendering issues
Previously, X window creation used an XVisualInfo object stored in
CStdWindow. This led to problems creating an OpenGL 3.x context, as
there is no way to get a framebuffer configuration from an XVisualInfo
object. As the reverse, however, is possible, this commit changes CStdWindow so
that a matching framebuffer configuration is stored - when a visual is needed,
it's allocated by glXGetVisualFromFBConfig.